Fixed bug in clisops.core.subset::_check_desc_coords()
If the dataset has a lat or lon dimension with only one value in it
then the corresponding `lat_bnds` or `lon_bnds` arguments were being
reversed. This was causing the selection to fail because no values
were matched.

I have added a simple check that will not reverse the selections
if the length of dimension is less than 2. There is a test added to
demonstrate this fix:

 tests/ops/test_subset/py::test_subset_with_lat_lon_single_values()